refactor(database): extract statistics repository from database manager
- Create new StatsRepository class to handle all statistical queries
- Move all time calculation and statistics methods from DatabaseManager to StatsRepository
- Update DatabaseManager to delegate statistics operations to StatsRepository
- Reduce DatabaseManager from 1446 lines to 129 lines by extracting logic
- Add comprehensive unit tests for the new StatsRepository
- Update progress tracking documentation to reflect completed refactoring phases
