+++ memory-bank/progress.md
@@ -5,7 +5,7 @@
 ## 当前状态
 
 - **活跃功能**: DatabaseManager 重构（拆分上帝类）
-- **当前阶段**: Phase 2 - 数据层拆分
+- **当前阶段**: Phase 4 - 完成
 
 ## 今日目标
 
@@ -13,6 +13,8 @@
 - [x] 创建 Design Doc
 - [x] Phase 1: 创建 ConnectionManager + MigrationManager
 - [x] Phase 2: 创建 SessionRepository
+- [x] Phase 3: 创建 StatsRepository
+- [x] Phase 4: 清理 DatabaseManager + 测试
 
 ## 最近完成
 
@@ -21,14 +23,14 @@
 - AGENTS.md 新增版本号更新规范
 - DatabaseManager 线程安全性分析
 - Decision-002 记录分析结果
-- Phase 1: ConnectionManager + MigrationManager (74+94 行)
-- Phase 2: SessionRepository (286 行) + 单元测试 (10 个测试用例)
-- DatabaseManager: 1446 → 767 行 (减少 47%)
+- Phase 1: ConnectionManager + MigrationManager (80+92 行)
+- Phase 2: SessionRepository (288 行) + 单元测试 (10 个测试用例)
+- Phase 3: StatsRepository (710 行) + 单元测试 (10 个测试用例)
+- DatabaseManager 重构完成: 1446 → 129 行 (减少 91%)
 
 ## 下一步
 
-- [ ] Phase 3: 创建 StatsRepository（提取统计查询）
-- [ ] Phase 4: 清理 DatabaseManager + 测试
+- [ ] 等待用户确认合并或进一步优化
 
 ## 笔记
